refactor: separate data fetching container element
